Move LogEntry to separate module
